+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-_No changes yet._
+<!-- _No changes yet._ -->
+
+- feat: add several new color themes
+- refactor: remove unused highlight color from theme
 
 ## v1.3.6 - 2026-02-16